ORGANIZE AND PROTOTYPE FUNCTIONS
Created and organized the file directory for dnn with a model and training procedure.
Added prototype functions for preprocessing the FFT computations, image processing,
model building and then finally training the model given the data generator.

Got rid of old file directories with data that should not be in the git repo.
Instead all data will be stored outside of this repo and should be referenced with
sys, or file location.