REFACTOR, EXP-SETUP CODE, PREPROCESSING, AUTOPEP8
Ran autopep8 on all python files.

Setup the preprocessing framework for the various inputs of the dnn models.
We have the ability to feed in data for a 1d-3d cnn type model and also
an auxiliary network based on the fragility maps.

Setup the HPC dir code to be run by experiments preplanned by me.
It helps that we use Singularity.

Also refactored code to be within the model, train, preprocessing modules.
Adding some initial tests for the module and should try to keep this updated.